requirements.dev.txt
mypy
flask-httpauth-stubs
python-lsp-server
python-lsp-ruff
pylsp-mypy
pytest
